# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for Library Management System",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "description": "Returns a list of all books with language support",
                "operationId": "getBooks",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference (en, ar)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to retrieve books"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Create a new book (Authentication required)",
                "operationId": "createBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "book": {
                                                    "$ref": "#/components/schemas/Book"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Book": {
                "description": "Book model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title_en": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "title_ar": {
                        "type": "string",
                        "example": "غاتسبي العظيم"
                    },
                    "description_en": {
                        "type": "string",
                        "example": "A classic American novel"
                    },
                    "description_ar": {
                        "type": "string",
                        "example": "رواية أمريكية كلاسيكية"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    }
                },
                "type": "object"
            },
            "Author": {
                "description": "Author model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "F. Scott Fitzgerald"
                    },
                    "bio": {
                        "type": "string",
                        "example": "American novelist"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreBookRequest": {
                "required": [
                    "title_en",
                    "title_ar",
                    "description_en",
                    "description_ar",
                    "price",
                    "author_id"
                ],
                "properties": {
                    "title_en": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "The Great Gatsby"
                    },
                    "title_ar": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "غاتسبي العظيم"
                    },
                    "description_en": {
                        "type": "string",
                        "example": "A classic American novel"
                    },
                    "description_ar": {
                        "type": "string",
                        "example": "رواية أمريكية كلاسيكية"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 29.99
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "BookResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Books retrieved successfully"
                    },
                    "data": {
                        "properties": {
                            "language": {
                                "type": "string",
                                "example": "en"
                            },
                            "books": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the token received from login endpoint",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}